+++ truck-geotrait/src/lib.rs
@@ -34,3 +34,5 @@ pub mod traits;
 pub use traits::*;
 /// Algorithms for curves and surfaces.
 pub mod algo;
+#[cfg(feature = "derive")]
+pub use truck_geoderive::*;