+++ truck-modeling/src/geometry.rs
@@ -4,11 +4,45 @@ use serde::{Deserialize, Serialize};
 #[doc(hidden)]
 pub use truck_geometry::{algo, inv_or_zero};
 pub use truck_geometry::{decorators::*, nurbs::*, specifieds::*};
-use truck_geotrait::{Invertible, ParametricSurface};
 pub use truck_polymesh::PolylineCurve;
 
+/// Leading curve for intersection
+#[derive(
+    Clone,
+    Debug,
+    Serialize,
+    Deserialize,
+    From,
+    ParametricCurve,
+    BoundedCurve,
+    ParameterDivision1D,
+    Cut,
+    Invertible,
+    SearchNearestParameterD1,
+    SearchParameterD1,
+)]
+pub enum Leader {
+    /// polyline curve
+    Polyline(PolylineCurve<Point3>),
+    /// bspline curve
+    BSpline(BSplineCurve<Point3>),
+}
+
 /// 3-dimensional curve
-#[derive(Clone, Debug, Serialize, Deserialize, From)]
+#[derive(
+    Clone,
+    Debug,
+    Serialize,
+    Deserialize,
+    From,
+    ParametricCurve,
+    BoundedCurve,
+    ParameterDivision1D,
+    Cut,
+    Invertible,
+    SearchNearestParameterD1,
+    SearchParameterD1,
+)]
 pub enum Curve {
     /// line
     Line(Line<Point3>),
@@ -17,7 +51,7 @@ pub enum Curve {
     /// 3-dimensional NURBS curve
     NURBSCurve(NURBSCurve<Vector4>),
     /// intersection curve
-    IntersectionCurve(IntersectionCurve<PolylineCurve<Point3>, Surface>),
+    IntersectionCurve(IntersectionCurve<Leader, Surface>),
 }
 
 macro_rules! derive_curve_method {
@@ -42,23 +76,19 @@ macro_rules! derive_curve_self_method {
     };
 }
 
-impl ParametricCurve for Curve {
-    type Point = Point3;
-    type Vector = Vector3;
-    fn subs(&self, t: f64) -> Self::Point { derive_curve_method!(self, ParametricCurve::subs, t) }
-    fn der(&self, t: f64) -> Self::Vector { derive_curve_method!(self, ParametricCurve::der, t) }
-    fn der2(&self, t: f64) -> Self::Vector { derive_curve_method!(self, ParametricCurve::der2, t) }
-}
-
-impl BoundedCurve for Curve {
-    fn parameter_range(&self) -> (f64, f64) {
-        derive_curve_method!(self, BoundedCurve::parameter_range,)
+impl Transformed<Matrix4> for Leader {
+    fn transform_by(&mut self, trans: Matrix4) {
+        match self {
+            Leader::Polyline(x) => Transformed::transform_by(x, trans),
+            Leader::BSpline(x) => Transformed::transform_by(x, trans),
+        }
     }
-}
-
-impl Invertible for Curve {
-    fn invert(&mut self) { derive_curve_method!(self, Invertible::invert,) }
-    fn inverse(&self) -> Self { derive_curve_self_method!(self, Invertible::inverse,) }
+    fn transformed(&self, trans: Matrix4) -> Leader {
+        match self {
+            Leader::Polyline(x) => Leader::Polyline(Transformed::transformed(x, trans)),
+            Leader::BSpline(x) => Leader::BSpline(Transformed::transformed(x, trans)),
+        }
+    }
 }
 
 impl Transformed<Matrix4> for Curve {
@@ -70,44 +100,9 @@ impl Transformed<Matrix4> for Curve {
     }
 }
 
-impl ParameterDivision1D for Curve {
-    type Point = Point3;
-    fn parameter_division(&self, range: (f64, f64), tol: f64) -> (Vec<f64>, Vec<Point3>) {
-        derive_curve_method!(self, ParameterDivision1D::parameter_division, range, tol)
-    }
-}
-
-impl Cut for Curve {
-    fn cut(&mut self, t: f64) -> Self { derive_curve_self_method!(self, Cut::cut, t) }
-}
-
-impl SearchNearestParameter<D1> for Curve {
-    type Point = Point3;
-    fn search_nearest_parameter<H: Into<SPHint1D>>(
-        &self,
-        point: Point3,
-        hint: H,
-        trials: usize,
-    ) -> Option<f64> {
-        derive_curve_method!(
-            self,
-            SearchNearestParameter::search_nearest_parameter,
-            point,
-            hint,
-            trials
-        )
-    }
-}
-
-impl SearchParameter<D1> for Curve {
-    type Point = Point3;
-    fn search_parameter<H: Into<SPHint1D>>(
-        &self,
-        point: Point3,
-        hint: H,
-        trials: usize,
-    ) -> Option<f64> {
-        derive_curve_method!(self, SearchParameter::search_parameter, point, hint, trials)
+impl From<IntersectionCurve<PolylineCurve<Point3>, Surface>> for Curve {
+    fn from(x: IntersectionCurve<PolylineCurve<Point3>, Surface>) -> Curve {
+        Curve::IntersectionCurve(x.change_leader(Leader::Polyline))
     }
 }
 
@@ -130,10 +125,39 @@ impl Curve {
             }
         }
     }
+    /// Make the leaders of `IntersectionCurve`s B-spline curves.
+    pub fn to_spline_leader(&mut self, p_tol: f64, d_tol: f64, trials: usize) -> bool {
+        if let Curve::IntersectionCurve(ref mut curve) = self {
+            if let Some(bspcurve) = BSplineCurve::cubic_approximation(
+                curve,
+                curve.parameter_range(),
+                p_tol,
+                d_tol,
+                trials,
+            ) {
+                *curve.leader_mut() = Leader::BSpline(bspcurve);
+                true
+            } else {
+                false
+            }
+        } else {
+            false
+        }
+    }
 }
 
 /// 3-dimensional surfaces
-#[derive(Clone, Debug, Serialize, Deserialize, From)]
+#[derive(
+    Clone,
+    Debug,
+    Serialize,
+    Deserialize,
+    From,
+    ParametricSurface,
+    ParameterDivision2D,
+    Invertible,
+    SearchParameterD2,
+)]
 pub enum Surface {
     /// Plane
     Plane(Plane),
@@ -167,45 +191,11 @@ macro_rules! derive_surface_self_method {
     };
 }
 
-impl ParametricSurface for Surface {
-    type Point = Point3;
-    type Vector = Vector3;
-    #[inline(always)]
-    fn subs(&self, u: f64, v: f64) -> Point3 {
-        derive_surface_method!(self, ParametricSurface::subs, u, v)
-    }
-    #[inline(always)]
-    fn uder(&self, u: f64, v: f64) -> Vector3 {
-        derive_surface_method!(self, ParametricSurface::uder, u, v)
-    }
-    #[inline(always)]
-    fn vder(&self, u: f64, v: f64) -> Vector3 {
-        derive_surface_method!(self, ParametricSurface::vder, u, v)
-    }
-    #[inline(always)]
-    fn uuder(&self, u: f64, v: f64) -> Vector3 {
-        derive_surface_method!(self, ParametricSurface::uuder, u, v)
-    }
-    #[inline(always)]
-    fn uvder(&self, u: f64, v: f64) -> Vector3 {
-        derive_surface_method!(self, ParametricSurface::uvder, u, v)
-    }
-    #[inline(always)]
-    fn vvder(&self, u: f64, v: f64) -> Vector3 {
-        derive_surface_method!(self, ParametricSurface::vvder, u, v)
-    }
-}
-
 impl ParametricSurface3D for Surface {
     #[inline(always)]
     fn normal(&self, u: f64, v: f64) -> Vector3 {
         derive_surface_method!(self, ParametricSurface3D::normal, u, v)
     }
-}
-
-impl Invertible for Surface {
-    fn invert(&mut self) { derive_surface_method!(self, Invertible::invert,) }
-    fn inverse(&self) -> Self { derive_surface_self_method!(self, Invertible::inverse,) }
 }
 
 impl Transformed<Matrix4> for Surface {
@@ -273,18 +263,6 @@ impl IncludeCurve<Curve> for Surface {
     }
 }
 
-impl SearchParameter<D2> for Surface {
-    type Point = Point3;
-    fn search_parameter<H: Into<SPHint2D>>(
-        &self,
-        point: Point3,
-        hint: H,
-        trials: usize,
-    ) -> Option<(f64, f64)> {
-        derive_surface_method!(self, SearchParameter::search_parameter, point, hint, trials)
-    }
-}
-
 impl SearchNearestParameter<D2> for Surface {
     type Point = Point3;
     fn search_nearest_parameter<H: Into<SPHint2D>>(
@@ -310,16 +288,5 @@ impl SearchNearestParameter<D2> for Surface {
                 algo::surface::search_nearest_parameter(rotted, point, hint, trials)
             }
         }
-    }
-}
-
-impl ParameterDivision2D for Surface {
-    #[inline(always)]
-    fn parameter_division(
-        &self,
-        range: ((f64, f64), (f64, f64)),
-        tol: f64,
-    ) -> (Vec<f64>, Vec<f64>) {
-        derive_surface_method!(self, ParameterDivision2D::parameter_division, range, tol)
     }
 }