+++ CHANGELOG.md
@@ -4,6 +4,10 @@ The version is of the bottom crate `truck-rendimpl`.
 
 ## Unreleased
 
+- Intersection curve with B-spline leader.
+- Derive macros for geometric traits.
+- Implement cubic B-spline approximation.
+- Fix clippy new lint.
 - Adds `cut_face_by_edge` to `Solid`.
 - Refactoring Search(Nearest)Parameter.
 - Converting express `trimmed_curve` to truck `TrimmedCurve`.